Added feature to login status to stay once login even when user goes to another link on page.
