+++ public/index.html
@@ -20,6 +20,7 @@
     <script defer src="signUpForm.js" type="module"></script>
     <script defer src="loginFeature.js" type="module"></script>
     <script defer src="signOut.js" type="module"></script>
+    <script defer src="closeButtons.js" type="module"></script>
